release: v0.1.10 prep - changelog, 100-line test split, childEnv cache/catalog pin fix
